+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.7.8
+- Added HttpStatus codes for generated ouput
+
+## 0.7.7
+- Changed the project structure to build seperate
+  binaries for CLI and Service.
+
+## 0.7.6
+- Updated the documentation for generate.
+
 ## 0.7.5
 - upgraded semantics version in build.gradle to support all
   EiffelEvents in the eiffel repo from topic-drop4 branch